+++ src/fastq.h
@@ -13,9 +13,6 @@
 #pragma GCC diagnostic ignored "-Wparentheses"
 
 #include <boost/iostreams/filtering_stream.hpp>
-#include <boost/iostreams/device/file.hpp>
-#include <boost/iostreams/filter/gzip.hpp>
-#include <boost/algorithm/string/predicate.hpp>
 
 #pragma GCC diagnostic pop
 
@@ -23,13 +20,14 @@ typedef boost::iostreams::filtering_stream<boost::iostreams::input>
   filter_stream_in;
 typedef boost::iostreams::filtering_stream<boost::iostreams::output>
   filter_stream_out;
-typedef std::string string;
 
 #define NO_LIMIT (std::numeric_limits<std::streamsize>::max())
 
 void open_fastx_in(filter_stream_in & instream, const std::string & infile);
 
 void open_fastx_out(filter_stream_out & outstream, const std::string & outfile, bool compress);
+
+void open_fastx_out(filter_stream_out & outstream, const std::string & outfile);
 
 // [[Rcpp::depends(BH)]]