+++ src/fastx_rename.cpp
@@ -1,5 +1,6 @@
 #include "fastq.h"
 
+
 //' @describeIn fastx_rename Rename reads in a FASTQ file
 //' @param infile (`character`) FAST(A/Q) file path, optionally gzipped.
 //' @param names (`character`) New read names.
@@ -20,7 +21,7 @@ std::string fastq_rename(
 
   std::size_t i = 0;
   while (instream.get()) {
-    string line;
+    std::string line;
     instream >> line;
     if (line.size() == 0) break;
     outstream << "@" << names[i++] << "\n";
@@ -41,27 +42,17 @@ std::string fasta_rename(
     std::string outfile
 ) {
   filter_stream_in instream;
-  if (boost::ends_with(infile, ".gz")) {
-    instream.push(boost::iostreams::gzip_decompressor());
-  }
-  instream.push(
-    boost::iostreams::file_source(infile, std::ios_base::in | std::ios_base::binary)
-  );
+  open_fastx_in(instream, infile);
 
   filter_stream_out outstream;
-  if (boost::ends_with(outfile, ".gz")) {
-    outstream.push(boost::iostreams::gzip_compressor());
-  }
-  outstream.push(
-    boost::iostreams::file_sink(outfile, std::ios_base::binary)
-  );
+  open_fastx_out(outstream, outfile, false);
   if (!outstream) {
     Rcpp::stop("Failed to open output file");
   }
 
   std::size_t i = 0;
   while (instream.get()) {
-    string line;
+    std::string line;
     instream >> line;
     if (line.size() == 0) break;
     outstream << ">" << names[i++] << "\n";